requirements.txt
binance-futures-connector
aiogram
pydantic
sqlalchemy - pre